Created EntityMaxHealthUpdateEvent, started with better logic for colored mobs.

+++ src/main/java/at/hannibal2/skyhanni/events/EntityHealthUpdateEvent.kt
@@ -2,4 +2,4 @@ package at.hannibal2.skyhanni.events
 
 import net.minecraft.entity.EntityLivingBase
 
-class EntityHealthUpdateEvent(val entity: EntityLivingBase, val health: Float) : LorenzEvent()
+class EntityHealthUpdateEvent(val entity: EntityLivingBase, val health: Int) : LorenzEvent()
